Create : Lists, Dicts, Loop, & Conditionals
The goal of this assignment is to demonstrate the ability to iterate over Python lists and dictionaries
along with utilizing conditional statements to conduct calculations.
Examples of these types of operations will be discussed in the book and lectures.
You must use ‘for loops’. To identify the largest and smallest numbers in the mpg list, you must use
conditional logic with an ‘if’ statement. You must use the variable names designated in the
assignment. Do not turn in broken code!